added direct transfer functionality to companion share and containers
If you press shift click on an item, instead of selecting the item for drag/drop, the item is now immediately transferred between the share windows.

added direct transfer functionality to companion share and containers

If you press shift click on an item, instead of selecting the item for drag/drop, the item is now immediately transferred between the share windows.

fixed direct transfer functionality to companion share and containers - a merge messed it up